requirements.txt

Pillow==10.2.0
moviepy==1.0.3
python-telegram-bot==20.3
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
python-magic==0.4.27
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.0.1
ebooklib==0.18
rarfile==4.0 